cs_ner_local/requirements.txt
pandas
openpyxl
openai
python-dotenv
nest_asyncio
# tiktoken  # 선택적: 정확한 토큰 카운팅 (없어도 동작함)
